some refactoring, add req/sub daemon shutdown, fix trie tests
