+++ node_tools/ctlr_data.py
@@ -1,7 +1,8 @@
 # coding: utf-8
 
 """
-    Default fpn ctlr state variables.
+    Default fpn ctlr state variables.  Trie keys must be in the set
+    `string.hexdigits`.
     :var net_trie: a Trie of JSON network/member state objects
     :var id_trie: a Trie of JSON member node net_id objects
 """
